requirements.txt
appdirs==1.4.3
astroid==2.4.0
boto3==1.12.46
botocore==1.15.46
certifi==2020.4.5.1
distlib==0.3.0
docutils==0.15.2
filelock==3.0.12
importlib-metadata==1.6.0
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
pipenv==2018.11.26
pylint==2.5.0
pypng==0.0.20
PyQRCode==1.2.1
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.18
virtualenv-clone==0.5.4
wrapt==1.12.1
zipp==3.1.0
